requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.46.0
packaging==24.1
pgvector~=0.3.4
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.6
whitenoise==6.7.0
